xrRender: Blenders: detail_still separated
