+++ src/Layers/xrRender/blenders/Blender_detail_still.h
@@ -1,9 +1,3 @@
-// Blender_Vertex_aref.h: interface for the CBlender_Vertex_aref class.
-//
-//////////////////////////////////////////////////////////////////////
-
-#if !defined(AFX_BLENDER_VERTEX_AREF_H__07141B30_A968_407E_86F8_D12702FE0B9B__INCLUDED_2)
-#define AFX_BLENDER_VERTEX_AREF_H__07141B30_A968_407E_86F8_D12702FE0B9B__INCLUDED_2
 #pragma once
 
 class CBlender_Detail_Still : public IBlender
@@ -12,15 +6,12 @@ public:
     xrP_BOOL oBlend;
 
 public:
-    virtual LPCSTR getComment() { return "LEVEL: detail objects"; }
-    virtual BOOL canBeLMAPped() { return FALSE; }
-    virtual void Save(IWriter& fs);
-    virtual void Load(IReader& fs, u16 version);
+    LPCSTR getComment() override;
+    void Save(IWriter& fs) override;
+    void Load(IReader& fs, u16 version) override;
 
-    virtual void Compile(CBlender_Compile& C);
+    void Compile(CBlender_Compile& C) override;
 
     CBlender_Detail_Still();
-    virtual ~CBlender_Detail_Still();
+    ~CBlender_Detail_Still() override = default;
 };
-
-#endif // !defined(AFX_BLENDER_VERTEX_AREF_H__07141B30_A968_407E_86F8_D12702FE0B9B__INCLUDED_2)
