+++ src/Layers/xrRenderPC_GL/xrRender_GL.vcxproj.filters
@@ -255,9 +255,6 @@
     <ClCompile Include="..\xrRender\R_Backend_hemi.cpp">
       <Filter>Refactored\Backend</Filter>
     </ClCompile>
-    <ClCompile Include="..\xrRender\blenders\Blender_detail_still.cpp">
-      <Filter>Shading templates\R1</Filter>
-    </ClCompile>
     <ClCompile Include="..\xrRender\blenders\Blender_Lm(EbB).cpp">
       <Filter>Shading templates\R1</Filter>
     </ClCompile>
@@ -681,6 +678,9 @@
     <ClCompile Include="..\xrRender\blenders\Blender_BmmD_deferred.cpp">
       <Filter>Shading templates</Filter>
     </ClCompile>
+    <ClCompile Include="..\xrRender\blenders\Blender_detail_still_deferred.cpp">
+      <Filter>Shading templates</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="stdafx.h">
@@ -783,9 +783,6 @@
       <Filter>Refactored\Backend</Filter>
     </ClInclude>
     <ClInclude Include="..\xrRender\blenders\Blender_tree.h">
-      <Filter>Shading templates\R1</Filter>
-    </ClInclude>
-    <ClInclude Include="..\xrRender\blenders\Blender_detail_still.h">
       <Filter>Shading templates\R1</Filter>
     </ClInclude>
     <ClInclude Include="..\xrRender\blenders\Blender_Lm(EbB).h">
@@ -1126,6 +1123,9 @@
     </ClInclude>
     <ClInclude Include="..\xrRender\r__pixel_calculator.h">
       <Filter>Core</Filter>
+    </ClInclude>
+    <ClInclude Include="..\xrRender\blenders\Blender_detail_still.h">
+      <Filter>Shading templates</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
