+++ src/Layers/xrRenderPC_R2/xrRender_R2.vcxproj
@@ -341,8 +341,8 @@
     <ClCompile Include="..\xrRender\Animation.cpp" />
     <ClCompile Include="..\xrRender\Blender.cpp" />
     <ClCompile Include="..\xrRender\blenders\Blender_BmmD_deferred.cpp" />
+    <ClCompile Include="..\xrRender\blenders\Blender_detail_still_deferred.cpp" />
     <ClCompile Include="..\xrRender\Blender_Recorder.cpp" />
-    <ClCompile Include="..\xrRender\blenders\Blender_detail_still.cpp" />
     <ClCompile Include="..\xrRender\blenders\Blender_Editor_Selection.cpp" />
     <ClCompile Include="..\xrRender\blenders\Blender_Editor_Wire.cpp" />
     <ClCompile Include="..\xrRender\blenders\Blender_Lm(EbB).cpp" />
