+++ src/Layers/xrRenderPC_R4/xrRender_R4.vcxproj
@@ -365,6 +365,7 @@
     <ClCompile Include="..\xrRenderDX10\3DFluid\dx103DFluidVolume.cpp" />
     <ClCompile Include="..\xrRenderDX10\Blender_Recorder_R3.cpp" />
     <ClCompile Include="..\xrRender\blenders\Blender_BmmD_deferred.cpp" />
+    <ClCompile Include="..\xrRender\blenders\Blender_detail_still_deferred.cpp" />
     <ClCompile Include="..\xrRender\blenders\dx10RainBlender.cpp" />
     <ClCompile Include="..\xrRenderDX10\dx10BufferUtils.cpp" />
     <ClCompile Include="..\xrRenderDX10\dx10ConstantBuffer.cpp" />
@@ -389,7 +390,6 @@
     <ClCompile Include="..\xrRender\Animation.cpp" />
     <ClCompile Include="..\xrRender\Blender.cpp" />
     <ClCompile Include="..\xrRender\Blender_Recorder.cpp" />
-    <ClCompile Include="..\xrRender\blenders\Blender_detail_still.cpp" />
     <ClCompile Include="..\xrRender\blenders\Blender_Editor_Selection.cpp" />
     <ClCompile Include="..\xrRender\blenders\Blender_Editor_Wire.cpp" />
     <ClCompile Include="..\xrRender\blenders\Blender_Lm(EbB).cpp" />
